# contracts/openapi/ibms-protected-api-v1.openapi.yaml
openapi: 3.0.3
info:
  title: VANTARIS IBMS Protected API
  version: ibms.contract.v1
  description: |
    Draft protected API boundary for VANTARIS IBMS.
    Sources: SECURITY-A6, A7, A8B, A10, CONTRACTS-A1, CORE-A0.
    Schemas marked as draft placeholders where runtime bodies are not fully confirmed.
servers:
  - url: /api
    description: Legacy unversioned backend prefix (current)
security:
  - bearerAuth: []
...
paths:
  /modeling/csv/list:
    get:
      tags: [Modeling]
      summary: List CSV files
      operationId: modelingCsvList
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          description: CSV file list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /modeling/csv/{device_code}:
    get:
      tags: [Modeling]
      summary: Get device CSV info
      operationId: modelingCsvInfo
      parameters:
        - $ref: '#/components/parameters/DeviceCode'
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          description: Device CSV metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /modeling/{device_code}/train:
    post:
      tags: [Modeling]
      summary: Train model
      operationId: modelingTrain
      parameters:
        - $ref: '#/components/parameters/DeviceCode'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Draft placeholder — training parameters vary by device
              additionalProperties: true
      responses:
        '202':
          description: Training accepted
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /modeling/{device_code}/predict:
    post:
      tags: [Modeling]
      summary: Single-point predict
      operationId: modelingPredict
      parameters:
        - $ref: '#/components/parameters/DeviceCode'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelingPredictRequest'
      responses:
        '200':
          description: Prediction result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /modeling/{device_code}/predict_future:
    post:
      tags: [Modeling]
      summary: Multi-day forecast
      operationId: modelingPredictFuture
      parameters:
        - $ref: '#/components/parameters/DeviceCode'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Draft placeholder — e.g. days, city, electricity_price
              additionalProperties: true
      responses:
        '200':
          description: Forecast result
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /modeling/{device_code}/model_info:
    get:
      tags: [Modeling]
      summary: Get model metadata
      operationId: modelingModelInfo
      parameters:
        - $ref: '#/components/parameters/DeviceCode'
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          description: Model info
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /modeling/{device_code}/{method}:
    get:
      tags: [Modeling]
      summary: Generic predictor GET dispatch
      operationId: modelingCallGet
      parameters:
        - $ref: '#/components/parameters/DeviceCode'
        - name: method
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          description: Method result
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [Modeling]
      summary: Generic predictor POST dispatch
      operationId: modelingCallPost
      parameters:
        - $ref: '#/components/parameters/DeviceCode'
        - name: method
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Method result
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /api/did/login — use placeholder token in docs only
  parameters:
    XTraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Optional correlation id for audit (SECURITY-A9)
    DeviceCode:
      name: device_code
      in: path
      required: true
      schema:
        type: string
      example: DEVICE_CODE_PLACEHOLDER
...
  schemas:
    ErrorResponse:
      type: object
      description: Existing Result.error style — draft
      properties:
        code:
          type: integer
        message:
          type: string
      additionalProperties: true
    SuccessResponse:
      type: object
      description: Existing Result.success style — draft
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          additionalProperties: true
      additionalProperties: true
...
    ModelingPredictRequest:
      type: object
      description: Draft placeholder schema for predict body
      additionalProperties: true
      example:
        timestamp: '2026-06-16T10:00:00'
        temperature: 24.5
...
  responses:
...
    UnauthorizedError:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Invalid token
    ForbiddenError:
      description: |
        Forbidden — valid JWT but missing required permission (A6B/A7B/A10B),
        or environment guard denied (e.g. SSE test in production, A8B).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 403
            message: 'Permission denied: requires [modeling:train]'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...